requirements.txt
fastapi==0.104.1
FastAPI_SQLAlchemy==0.2.1
SQLAlchemy==2.0.25
asyncpg==0.30.0
uvicorn[standard]==0.21.1
beanie==1.23.6
pyjwt[crypto]==2.9.0
httpx==0.24.1
aiohttp==3.11.10
slowapi==0.1.8
python-multipart==0.0.6
tomli==2.0.1
pydantic-settings==2.1.0
aioredis==1.3.1
alembic==1.13.2
asgi-correlation-id==4.3.4
pydub==0.25.1
